pyproject.toml
[project]
name = "ai-poc-branding-extraction"
version = "0.1.0"
description = "AI agent for extracting branding elements from websites using Google ADK + OpenAI"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cssutils>=2.11.1",
    "fastapi>=0.135.2",
    "google-adk[litellm]>=1.28.0",
    "httpx>=0.28.1",
    "litellm>=1.82.6",
    "lxml>=6.0.2",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "uvicorn[standard]>=0.42.0",
]
